[C] Fixed transition between Stats and new game.
Added a setTimeout to prevent the user from triggering the PLAY AGAIN
button before the transition between the game end and the Stats screen
finished. When that happened, there wouldn't be any visible screen and
the user would get an empty page.
Also tweaked the .transition-out rule to keep the screen stuck to the
top of the page while it's leaving